requirements.txt
# covid19co_tweet requirements
pandas
numpy
unidecode
jupyterlab
kaggle
tweepy
pylint
autopep8
python-dotenv
